+++ frontend/.eslintcache
@@ -1,1 +1,1 @@
-[{"D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\App.js":"1","D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\mainSection\\index.jsx":"2","D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\nav\\index.jsx":"3","D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\shortcuts\\index.jsx":"4","D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\index.js":"5","D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\svgbackground\\index.js":"6","D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\mobileNav\\index.jsx":"7","D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\navLinks\\index.jsx":"8"},{"size":380,"mtime":1610477109708,"results":"9","hashOfConfig":"10"},{"size":589,"mtime":1610477129463,"results":"11","hashOfConfig":"10"},{"size":528,"mtime":1610476541145,"results":"12","hashOfConfig":"10"},{"size":877,"mtime":1610474918546,"results":"13","hashOfConfig":"10"},{"size":517,"mtime":1609056805508,"results":"14","hashOfConfig":"10"},{"size":894,"mtime":1610477066839,"results":"15","hashOfConfig":"10"},{"size":1189,"mtime":1610476546784,"results":"16","hashOfConfig":"10"},{"size":454,"mtime":1610475498323,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"w9tfkj",{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},"D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\App.js",[],"D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\mainSection\\index.jsx",[],"D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\nav\\index.jsx",[],"D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\shortcuts\\index.jsx",[],"D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\index.js",[],["35","36"],"D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\svgbackground\\index.js",[],"D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\mobileNav\\index.jsx",[],"D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\navLinks\\index.jsx",[],{"ruleId":"37","replacedBy":"38"},{"ruleId":"39","replacedBy":"40"},"no-native-reassign",["41"],"no-negated-in-lhs",["42"],"no-global-assign","no-unsafe-negation"]
+[{"D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\App.js":"1","D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\shortcuts\\index.jsx":"2","D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\index.js":"3","D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\svgbackground\\index.js":"4","D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\navLinks\\index.jsx":"5","D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\homeNav\\index.jsx":"6","D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\homePage\\index.jsx":"7","D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\Nav\\index.jsx":"8"},{"size":333,"mtime":1610610450040,"results":"9","hashOfConfig":"10"},{"size":877,"mtime":1610525089845,"results":"11","hashOfConfig":"10"},{"size":517,"mtime":1609056805508,"results":"12","hashOfConfig":"10"},{"size":894,"mtime":1610477066839,"results":"13","hashOfConfig":"10"},{"size":454,"mtime":1610475498323,"results":"14","hashOfConfig":"10"},{"size":534,"mtime":1610610616835,"results":"15","hashOfConfig":"10"},{"size":710,"mtime":1610610897667,"results":"16","hashOfConfig":"10"},{"size":1177,"mtime":1610610965351,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"w9tfkj",{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\App.js",[],"D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\shortcuts\\index.jsx",[],"D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\index.js",[],["35","36"],"D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\svgbackground\\index.js",[],"D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\navLinks\\index.jsx",[],"D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\homeNav\\index.jsx",[],"D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\homePage\\index.jsx",[],"D:\\code\\web\\bitbyteGithub\\TPC-website\\frontend\\src\\components\\Nav\\index.jsx",[],{"ruleId":"37","replacedBy":"38"},{"ruleId":"39","replacedBy":"40"},"no-native-reassign",["41"],"no-negated-in-lhs",["42"],"no-global-assign","no-unsafe-negation"]
