+++ frontend/src/App.js
@@ -1,16 +1,14 @@
 import './App.css';
-import Main from "./components/mainSection";
-import Nav from "./components/nav";
+import HomePage from "./components/homePage";
 import Shortcuts from "./components/shortcuts";
-import MobileNav from "./components/mobileNav";
+import Nav from "./components/Nav";
 
 function App() {
   return (
     <div className="App">
       <Shortcuts />
-      <MobileNav />
-      <Main />
       <Nav />
+      <HomePage />
 
     </div>
   );
