+++ frontend/src/components/homePage/style.module.css
@@ -1,3 +1,10 @@
+.home {
+  grid-area: primary;
+  min-height: 100vh;
+  display: grid;
+  grid-template-columns: 66fr 33fr;
+  grid-template-areas: "main nav";
+}
 .container {
   grid-area: main;
   display: flex;
@@ -32,6 +39,13 @@
 }
 
 @media only screen and (max-width: 1080px) {
+  .home {
+    grid-template-columns: auto;
+    grid-template-rows: 50% 40%;
+    grid-template-areas:
+      "main"
+      "footer";
+  }
   .container {
     justify-content: right;
   }
