--- frontend/src/components/mobileNav/style.module.css
@@ -1,42 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-}
-.container {
-  display: none;
-}
-
-@media only screen and (max-width: 1080px) {
-  .container {
-    display: block;
-    grid-area: nav;
-  }
-  svg {
-    position: fixed;
-    left: 0;
-    margin: 2vh;
-    z-index: 4;
-    cursor: pointer;
-  }
-
-  .navContainer {
-    position: fixed;
-    top: 0;
-    background: #14335f;
-    height: 100vh;
-    width: 100vw;
-    z-index: 4;
-    color: white;
-    list-style: none;
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-    clip-path: circle(66px at 0 0);
-    -webkit-clip-path: circle(66px at 0 0);
-    transition: all 0.5s ease-out;
-  }
-  .activeNav {
-    clip-path: circle(200vh at 0 0);
-    -webkit-clip-path: circle(200vh at 0 0);
-  }
-}
