+++ frontend/src/components/nav/style.module.css
@@ -1,43 +1,42 @@
+* {
+  margin: 0;
+  padding: 0;
+}
 .container {
-  padding: 3vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-around;
-  z-index: 3;
-}
-.buttons {
-  background: var(--nav-btn-background);
-  padding: 1vh 2vh;
-  border-radius: 20px;
-}
-.navBtn {
-  margin: 0 2vh;
-  color: white;
-  padding: 1vh;
-  cursor: pointer;
-}
-.navBtn:hover {
-  color: rgb(241, 238, 238);
-}
-.nav {
-  list-style: none;
-  margin-top: 2vh;
-}
-
-.vector {
-  width: 100%;
+  display: none;
 }
 
 @media only screen and (max-width: 1080px) {
-  .buttons,
-  .nav {
-    display: none;
-  }
   .container {
-    grid-area: footer;
+    display: block;
+    grid-area: nav;
   }
-  .vector {
-    width: 80%;
+  svg {
+    position: fixed;
+    left: 0;
+    margin: 2vh;
+    z-index: 4;
+    cursor: pointer;
+  }
+
+  .navContainer {
+    position: fixed;
+    top: 0;
+    background: #14335f;
+    height: 100vh;
+    width: 100vw;
+    z-index: 4;
+    color: white;
+    list-style: none;
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    clip-path: circle(66px at 0 0);
+    -webkit-clip-path: circle(66px at 0 0);
+    transition: all 0.5s ease-out;
+  }
+  .activeNav {
+    clip-path: circle(200vh at 0 0);
+    -webkit-clip-path: circle(200vh at 0 0);
   }
 }
